feat: 可选的 AsyncLogger

+++ sese/src/net/http/V2Http2Server.cpp
@@ -540,7 +540,7 @@ void Http2Server::sendHeader(Http2Context &ctx,
             info.flags = FRAME_FLAG_END_HEADERS;
         }
 
-        info.length = req;
+        info.length = (uint32_t) req;
         writeFrame(conn, info);
         conn->context.write(buffer, req);
